requirements.txt
python-telegram-bot==20.3
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.95.0
pytz==2023.3
python-dotenv==1.0.0
cachetools==5.3.1